# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Chat Service API
  version: 1.0.0
  description: API for the chat application backend service.
servers:
  - url: http://localhost:8080
paths:
  /messages:
    get:
      tags:
        - messages
      summary: Retrieve message history
      description: Returns a list of all messages in the chat history.
      operationId: getMessages
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of messages.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
    post:
      tags:
        - messages
      summary: Send a new message
      description: Enqueues a new chat message to the Kafka topic. Persistence and broadcast occur asynchronously.
      operationId: sendMessage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  description: The ID of the user sending the message.
                content:
                  type: string
                  description: The message text.
      responses:
        '202':
          description: Message accepted and enqueued for processing.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message_id:
                    type: string
                    description: Server-assigned message identifier.
                  status:
                    type: string
                    example: enqueued
        '400':
          description: Invalid request body.
components:
  schemas:
    Message:
      type: object
      properties:
        message_id:
          type: string
          description: Unique identifier for the message.
        user_id:
          type: string
          description: The ID of the user who sent the message.
        content:
          type: string
          description: The message text.
        timestamp:
          type: string
          format: date-time
          description: The timestamp when the message was created.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT